requirements.txt
wikipedia~=1.4.0
python-dotenv~=1.0.0
aiogram~=2.25.1